Revert yarn, fewer diagnostics, clear cache
Using an older version of yarn did not help.  It is definitely the
first invocation of yarn that is changing the lockfile, even with
--frozen-lockfile.  It is hard to tell if that option even does
anything since providing --frzn-lf does not error and seems to have
the same behavior, as does providing --asdfsdfd

Clear cache on Travis CI, although we are not caching the yarn cache.
Provide information about the yarn cache.