Update to rug-functions-travis 0.18.0
Get rid of call to GitHub API, relying on travis-restart-build to do
all that.  Update tests.

Update TravisBuilds event handler, create restart button with proper
arguments and use ChatId.screenName rather than .id when sending a DM,
and only attempt to send a DM is there is a screen name to send it to.
Add tests for the TravisBuilds event handler.

Fix GitHub token for Travis CI build.